new user-agent string for storage record API reqs
