+++ packages/attest/src/internal/package-version.cjs
@@ -3,5 +3,5 @@
 // ESLint rules and doesn't work reliably across all Node.js versions.
 // By keeping this as a .cjs file, we can use require() naturally and export
 // the version for the ESM modules to import.
-const packageJson = require('../package.json')
+const packageJson = require('../../package.json')
 module.exports = {version: packageJson.version}
