+++ packages/attest/src/store.ts
@@ -1,7 +1,7 @@
 import * as github from '@actions/github'
 import {retry} from '@octokit/plugin-retry'
 import {RequestHeaders} from '@octokit/types'
-import {version} from './package-version.cjs'
+import {getUserAgent} from './internal/utils.js'
 
 const CREATE_ATTESTATION_REQUEST = 'POST /repos/{owner}/{repo}/attestations'
 const DEFAULT_RETRY_COUNT = 5
@@ -51,18 +51,4 @@ export const writeAttestation = async (
     const message = err instanceof Error ? err.message : err
     throw new Error(`Failed to persist attestation: ${message}`)
   }
-}
-
-const getUserAgent = (): string => {
-  const baseUserAgent = `@actions/attest-${version}`
-
-  const orchId = process.env['ACTIONS_ORCHESTRATION_ID']
-  if (orchId) {
-    // Sanitize the orchestration ID to ensure it contains only valid characters
-    // Valid characters: 0-9, a-z, _, -, .
-    const sanitizedId = orchId.replace(/[^a-z0-9_.-]/gi, '_')
-    return `${baseUserAgent} actions_orchestration_id/${sanitizedId}`
-  }
-
-  return baseUserAgent
 }
